+++ modules/bx_detector_monitor.hh
@@ -35,74 +35,74 @@ class bx_detector_monitor: public bx_base_module {
   private:
      
       // occupancy
-    int total_triggers;
-    int n_triggers [4]; 
+    int32_t total_triggers;
+    int32_t n_triggers [4]; 
   
-    int check_after_precalib_done;
+    int32_t check_after_precalib_done;
 
-    int calibration_mode;
-    int disabling_lg;
+    int32_t calibration_mode;
+    int32_t disabling_lg;
 
-    int n[5];
-    int n_on[5];
+    int32_t n[5];
+    int32_t n_on[5];
 
-    int start_lg_off;
-    int start_lg_on;
-    int end_lg_off;
-    int end_lg_on;
-    int start_lg_off_charge;
-    int start_lg_on_charge;
-    int end_lg_off_charge;
-    int end_lg_on_charge;
+    int32_t start_lg_off;
+    int32_t start_lg_on;
+    int32_t end_lg_off;
+    int32_t end_lg_on;
+    int32_t start_lg_off_charge;
+    int32_t start_lg_on_charge;
+    int32_t end_lg_off_charge;
+    int32_t end_lg_on_charge;
 
-    std::vector<int> times_found_noisy[5][4];
-    std::vector<int> times_found_empty[5][4];
+    std::vector<int32_t> times_found_noisy[5][4];
+    std::vector<int32_t> times_found_empty[5][4];
 
       //precalibration check
     std::vector<float> nhits_in_precalib;
       
     //cumulative_off (in any trigger)
-    std::vector<int> cummulative_off_lg;
-    std::vector<int> cummulative_charge_off_lg;
+    std::vector<int32_t> cummulative_off_lg;
+    std::vector<int32_t> cummulative_charge_off_lg;
 
-    std::vector<int> latest_off_lg_p;
-    std::vector<int> latest_off_lg_HV;
+    std::vector<int32_t> latest_off_lg_p;
+    std::vector<int32_t> latest_off_lg_HV;
     std::vector<double> cummulative_occupancy_p;
     std::vector<double> cummulative_lbnb_occupancy_p;
     std::vector<double> cummulative_occupancy_HV;
     std::vector<double> cummulative_hvb_occupancy_HV;
 
       //how many time the lg changed its status to off (on) in pulser and HV triggers
-    std::vector<int> lg_Nx_to_off_p;
-    std::vector<int> lg_Nx_to_on_p;
-    std::vector<int> lg_Nx_to_off_HV;
-    std::vector<int> lg_Nx_to_on_HV;
+    std::vector<int32_t> lg_Nx_to_off_p;
+    std::vector<int32_t> lg_Nx_to_on_p;
+    std::vector<int32_t> lg_Nx_to_off_HV;
+    std::vector<int32_t> lg_Nx_to_on_HV;
 
       //the event number of the LATEST change
-    std::vector<int> ev_lg_to_off_p;
-    std::vector<int> ev_lg_to_on_p;
-    std::vector<int> ev_lg_to_off_HV;
-    std::vector<int> ev_lg_to_on_HV;
+    std::vector<int32_t> ev_lg_to_off_p;
+    std::vector<int32_t> ev_lg_to_on_p;
+    std::vector<int32_t> ev_lg_to_off_HV;
+    std::vector<int32_t> ev_lg_to_on_HV;
 
       // for neutrino trigger rate and nhits rate
-    unsigned long n_very_first_event_tsec;
-    int n_trgrate_cycle;
-    int n_events_in_this_cycle;
+    uint32_t n_very_first_event_tsec;
+    int32_t n_trgrate_cycle;
+    int32_t n_events_in_this_cycle;
     double n_time_last_previous_cycle;
-    int n_nhits_in_cycle;
+    int32_t n_nhits_in_cycle;
 
     std::list<double> tlist;
-    std::list<int> neutrino_nhits_list;
+    std::list<int32_t> neutrino_nhits_list;
     
       // for random nhits rate
-    unsigned long r_very_first_event_tsec;
-    int r_trgrate_cycle;
-    int r_events_in_this_cycle;
+    uint32_t r_very_first_event_tsec;
+    int32_t r_trgrate_cycle;
+    int32_t r_events_in_this_cycle;
     double r_time_last_previous_cycle;
-    int r_nhits_in_cycle;
+    int32_t r_nhits_in_cycle;
 
     std::list<double> random_tlist;
-    std::list<int> random_nhits_list;
+    std::list<int32_t> random_nhits_list;
 
      
     //pulser
@@ -126,8 +126,8 @@ class bx_detector_monitor: public bx_base_module {
     TH1F* pulser_nhits;    
 
 
-    int pulser_bins;
-    int max_events;
+    int32_t pulser_bins;
+    int32_t max_events;
   
     enum module_type{
       crate = 0,
@@ -144,8 +144,8 @@ class bx_detector_monitor: public bx_base_module {
       laser = 3,
    };
 
-  std::cmap <int, std::string> module_type_map;
-  std::cmap <int, std::string> trigger_type_map;
+  std::cmap <int32_t, std::string> module_type_map;
+  std::cmap <int32_t, std::string> trigger_type_map;
 
 };
 #endif
